refactor: 강의목록 관련 파일 폴더 구조 정리
