+++ src/domains/course/components/courseList/courseCard/CourseCheckBox.tsx
